Move state function to liveview where it is actually used
