+++ app/lib/photo_tagger_web/controllers/photo_controller.ex
@@ -4,7 +4,6 @@ defmodule PhotoTaggerWeb.PhotoController do
   alias PhotoTagger.Repo
   alias PhotoTagger.Gallery
   alias PhotoTagger.Gallery.Photo
-  require Logger
 
   def build_url(folder, photo, tags, tail \\ nil) do
     uri = URI.new!("/")
@@ -16,47 +15,6 @@ defmodule PhotoTaggerWeb.PhotoController do
     uri = if(!Enum.empty?(tags), do: URI.append_query(uri, query), else: uri)
 
     URI.to_string(uri)
-  end
-
-  def expand_state(%{folder: folder, tags: tags, photo_id: photo_id}) do
-    filtered_photos =
-      case {folder, tags} do
-        {nil, []} -> Gallery.list_photos()
-        {nil, tags} -> Gallery.list_photos_by_all_tags(tags)
-        {folder, []} -> Gallery.list_photos_by_folder(folder)
-        {folder, tags} -> Gallery.list_photos_by_folder_and_tags(folder, tags)
-      end
-    filtered_photos = Repo.preload(filtered_photos, :tags)
-
-    photo =
-      case photo_id do
-        nil ->
-          nil
-
-        id ->
-          p = Gallery.get_photo!(id)
-          Repo.preload(p, :tags)
-      end
-
-    recommended_tags =
-      Enum.reduce(filtered_photos, MapSet.new(), fn photo, acc ->
-        MapSet.union(acc, MapSet.new(photo.tags))
-      end)
-      |> MapSet.to_list()
-      |> Enum.sort_by(& &1.name)
-
-    all_folders = Gallery.list_folders_include_tags()
-    all_tags = Gallery.list_tags()
-
-    %{
-      folder: folder,
-      all_folders: all_folders,
-      tags: tags,
-      all_tags: all_tags,
-      photo: photo,
-      filtered_photos: filtered_photos,
-      recommended_tags: recommended_tags
-    }
   end
 
   def new(conn, _params) do
